requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
openai>=1.0.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# HTML parsing for LinkedIn profiles
beautifulsoup4==4.12.2

# Data processing
pandas==2.1.4
python-dateutil==2.8.2

# HTTP client
httpx==0.25.2

# Salesforce integration
simple-salesforce==1.12.5

# FastAPI backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23

# RAG/Vector Database
chromadb==1.3.3
sentence-transformers==5.1.2
chroma-hnswlib==0.7.6

# Google Gemini SDK
google-genai>=0.2.0
